refactor(workflow): house WorkflowEngineError in workflow-errors
Move the engine domain error class next to other workflow failures and
add result_too_large for upcoming oversized-return policy. Re-export
from workflow-api for existing imports.
